requirements.txt
yt-dlp
opencv-python-headless
numpy
requests
Pillow
